+++ src/azure-cli/azure/cli/command_modules/aro/aaz/latest/identity/_create.py
@@ -10,6 +10,7 @@
 
 from azure.cli.core.aaz import *
 
+
 class Create(AAZCommand):
     """Create an identity in the specified subscription and resource group.
     """