+++ src/azure-cli/azure/cli/command_modules/vm/_params.py
@@ -79,6 +79,7 @@ def load_arguments(self, _):
     PublicNetworkAccess = ['Disabled', 'Enabled']
     RepairAction = ['Replace', 'Restart', 'Reimage']
     ReplicationMode = ['Full', 'Shallow']
+    ReservationType = ['Block', 'Open', 'Targeted']
     SecurityEncryptionType = ['VMGuestStateOnly', 'DiskWithVMGuestState', 'NonPersistedTPM']
     SecurityType = ['TrustedLaunch', 'Standard', 'ConfidentialVM']
     SnapshotStorageAccountTypes = ['Premium_LRS', 'Standard_LRS', 'Standard_ZRS']
@@ -502,6 +503,7 @@ def load_arguments(self, _):
             c.argument('key_incarnation_id', type=int, help='Increase the value of this property allows user to reset the key used for securing communication channel between guest and host.')
             c.argument('security_type', arg_type=get_enum_type(SecurityType, default=None), help='Specify the security type of the virtual machine.')
             c.argument('zone_movement', arg_type=get_three_state_flag(), help='Indicates if zone movement is enabled. By default isEnabled is set to false i.e VM can\'t be moved from one zone to another.')
+            c.argument('disable_capacity_reservation_assignment', options_list=['--disable-capacity-reservation-assignment', '--no-cap-reservation'], arg_type=get_three_state_flag(), help='Explicitly opt out the VM from being associated with any capacity reservation. The VM will consume publicly available capacity.')
 
     with self.argument_context('vm create', arg_group='Storage') as c:
         c.argument('attach_os_disk', help='Attach an existing OS disk to the VM. Can use the name or ID of a managed disk or the URI to an unmanaged disk VHD.')
@@ -987,6 +989,9 @@ def load_arguments(self, _):
             c.argument('enable_automatic_zone_balancing', arg_type=get_three_state_flag(), options_list=['--enable-automatic-zone-balancing', '--enable-zone-balancing'], help='Specify whether automatic AZ balancing should be enabled on the virtualmachine scale set.')
             c.argument('automatic_zone_balancing_strategy', arg_type=get_enum_type(AutomaticZoneBalancingStrategy), options_list=['--automatic-zone-balancing-strategy', '--balancing-strategy'], help='Type of rebalance strategy that will be used for rebalancing virtualmachines in the scale set across availability zones.')
             c.argument('automatic_zone_balancing_behavior', arg_type=get_enum_type(AutomaticZoneBalancingBehavior), options_list=['--automatic-zone-balancing-behavior', '--balancing-behavior'], help='Type of rebalance behavior that will be used for recreating virtualmachines in the scale set across availability zones.')
+            c.argument('disable_capacity_reservation_assignment', arg_type=get_three_state_flag(),
+                       options_list=['--disable-capacity-reservation-assignment', '--no-cap-reservation'],
+                       help='Specifies whether the virtual machine scale set is explicitly opted out from being associated with any capacity reservation. When set to true, its virtual machines will not be allowed to implicitly or explicitly associate with any type of capacity reservation and will consume publicly available capacity. This option applies to the parent VMSS only.')
 
     with self.argument_context('vmss update') as c:
         c.argument('instance_id', id_part='child_name_1', help="Update the VM instance with this ID. If missing, update the VMSS.")
@@ -1645,6 +1650,7 @@ def load_arguments(self, _):
                    help='The name of the capacity reservation group.')
         c.argument('tags', tags_type)
         c.argument('sharing_profile', nargs='*', help='Space-separated subscription resource IDs or nothing. Specify the settings to enable sharing across subscriptions for the capacity reservation group resource. Specify it to nothing to unsharing.')
+        c.argument('reservation_type', arg_type=get_enum_type(ReservationType), help='The capacity reservation type. The reservation type cannot be changed after the capacity reservation group is created.')
 
     with self.argument_context('capacity reservation group create') as c:
         c.argument('zones', zones_type, help='Availability Zones to use for this capacity reservation group. If not provided, the group supports only regional resources in the region. If provided, enforces each capacity reservation in the group to be in one of the zones.')