+++ src/azure-cli/azure/cli/command_modules/vm/tests/latest/test_vm_commands.py
@@ -9255,7 +9255,7 @@ class VMGalleryImage(ScenarioTest):
         ])
 
     @AllowLargeResponse(size_kb=99999)
-    @ResourceGroupPreparer(location='eastus2')
+    @ResourceGroupPreparer(location='centralus')
     def test_create_vm_with_community_gallery_image(self, resource_group, resource_group_location):
         self.kwargs.update({
             'vm': self.create_random_name('vm', 16),
@@ -9289,13 +9289,13 @@ class VMGalleryImage(ScenarioTest):
         self.cmd('sig image-version create -g {rg} --gallery-name {gallery} --gallery-image-definition {image} --gallery-image-version {version} --managed-image {captured} --replica-count 1')
         self.kwargs['public_name'] = self.cmd('sig show --gallery-name {gallery} --resource-group {rg} --select Permissions').get_output_in_json()['sharingProfile']['communityGalleryInfo']['publicNames'][0]
 
-        self.cmd('sig image-version show-community --gallery-image-definition {image} --public-gallery-name {public_name} -l eastus2 --gallery-image-version {version}',
+        self.cmd('sig image-version show-community --gallery-image-definition {image} --public-gallery-name {public_name} -l centralus --gallery-image-version {version}',
             checks=[
                 self.check('name', '{version}'),
                 self.check('uniqueId', '/CommunityGalleries/{public_name}/Images/{image}/Versions/{version}')
             ])
 
-        self.kwargs['community_gallery_image_version'] = self.cmd('sig image-version show-community --gallery-image-definition {image} --public-gallery-name {public_name} --location eastus2 --gallery-image-version {version}').get_output_in_json()['uniqueId']
+        self.kwargs['community_gallery_image_version'] = self.cmd('sig image-version show-community --gallery-image-definition {image} --public-gallery-name {public_name} --location centralus --gallery-image-version {version}').get_output_in_json()['uniqueId']
 
         self.cmd('vm create -g {rg} -n {vm_with_community_gallery} --image {community_gallery_image_version} --size Standard_D2s_v3 '
                  '--admin-username gallerytest --generate-ssh-keys --nsg-rule None --accept-term --subnet {subnet} --vnet-name {vnet}')
@@ -12877,7 +12877,7 @@ class VMTrustedLaunchScenarioTest(ScenarioTest):
         ])
 
     @AllowLargeResponse(size_kb=99999)
-    @ResourceGroupPreparer(name_prefix='cli_vm_vmss_proxy_agent_control_profile_reference', location='eastus2')
+    @ResourceGroupPreparer(name_prefix='cli_vm_vmss_proxy_agent_control_profile_reference', location='centralus')
     def test_vm_vmss_proxy_agent_control_profile_reference(self, resource_group):
         self.kwargs.update({
             'nsg1': self.create_random_name('nsg', 10),
@@ -13341,6 +13341,71 @@ class CapacityReservationScenarioTest(ScenarioTest):
         if self.is_live:
             time.sleep(60)
         self.cmd('capacity reservation group delete -n {reservation_group} -g {rg} --yes')
+
+    # Open Capacity Reservation is currently enabled only in the East US 2 EUAP Canary region.
+    # Provide private package for service team to test it out.
+    @live_only()
+    @ResourceGroupPreparer(name_prefix='cli_test_open_capacity_reservation_', location='eastus2euap')
+    def test_open_capacity_reservation(self, resource_group):
+
+        self.kwargs.update({
+            'rg': resource_group,
+            'location': 'eastus2euap',
+            'reservation_group': self.create_random_name('cli_open_reservation_group_', 40),
+            'reservation_name': self.create_random_name('cli_open_reservation_', 40),
+            'vm': self.create_random_name('cli-open-vm-', 20),
+            'sku': 'Standard_DS1_v2',
+            'image': 'Canonical:UbuntuServer:18.04-LTS:latest',
+            'ssh_key': TEST_SSH_KEY_PUB,
+        })
+
+        self.cmd('capacity reservation group create -n {reservation_group} -g {rg} -l {location} '
+                 '--reservation-type Open',
+                 checks=[
+                     self.check('name', '{reservation_group}'),
+                     self.check('reservationType', 'Open')
+                 ])
+
+        self.cmd('capacity reservation create -c {reservation_group} -n {reservation_name} -g {rg} '
+                 '--sku {sku} --capacity 1',
+                 checks=[
+                     self.check('name', '{reservation_name}'),
+                     self.check('sku.name', '{sku}'),
+                     self.check('sku.capacity', 1),
+                     self.check('provisioningState', 'Succeeded')
+                 ])
+
+        self.cmd('vm create -g {rg} -n {vm} -l {location} --image {image} --size {sku} '
+                 '--admin-username azureuser '
+                 '--ssh-key-value "{ssh_key}" --public-ip-address "" --nsg-rule None --no-cap-reservation true')
+
+        self.cmd('vm show -g {rg} -n {vm}', checks=[
+            self.check('provisioningState', 'Succeeded'),
+            self.check('capacityReservation.disableCapacityReservationAssignment', True)
+        ])
+
+        self.cmd('vm deallocate -g {rg} -n {vm}')
+        self.cmd('vm update -g {rg} -n {vm} --disable-capacity-reservation-assignment false', checks=[
+            self.check('capacityReservation.disableCapacityReservationAssignment', False)
+        ])
+        self.cmd('vm start -g {rg} -n {vm}')
+
+        self.cmd('vm get-instance-view -g {rg} -n {vm}', checks=[
+            self.check('instanceView.capacityReservationType', 'Open')
+        ])
+
+        utilization = \
+            self.cmd('capacity reservation show -c {reservation_group} -n {reservation_name} -g {rg} '
+                     '--instance-view', checks=[
+                self.exists('instanceView.utilizationInfo.usedReservedCountBySubscription')
+            ]).get_output_in_json()['instanceView']['utilizationInfo']['usedReservedCountBySubscription']
+        self.assertEqual(1, len(utilization))
+        self.assertEqual(1, next(iter(utilization.values())))
+
+        self.cmd('vm deallocate -g {rg} -n {vm}')
+        self.cmd('vm update -g {rg} -n {vm} --disable-capacity-reservation-assignment true', checks=[
+            self.check('capacityReservation.disableCapacityReservationAssignment', True)
+        ])
 
     @record_only()  # Some special subscriptions can test it.
     @ResourceGroupPreparer(name_prefix='cli_test_capacity_reservation_sharing_profile', location='westEurope')